Remove UAA from Console Helm chart

+++ deploy/db/scripts/run-postflight-job.k8s.sh
@@ -66,6 +66,10 @@ echo "Database operation(s) complete."
 
 
 # Remove the lock file on the shared volume
+# Check if Upgrade Lock file exists
+if [ ! -f "/$UPGRADE_VOLUME/$UPGRADE_LOCK_FILENAME" ]; then
+  exit 1
+fi
 echo "Removing the $UPGRADE_LOCK_FILENAME file from the shared upgrade volume $UPGRADE_VOLUME."
 rm /$UPGRADE_VOLUME/$UPGRADE_LOCK_FILENAME || true
